application: Split application.cc into multiple files
application.cc is one of the slowest files to compile. However the
bigger problem is that it has so many dependencies which effectively
puts it at the very end of the compile graph. In practice this means that
most of the redpanda builds are effectively single threaded in the end and
spend a minute or two just with one core being busy compiling
application.cc (this can be even worse on machines with low single core
perf).

To counteract that this PR splits up application.cc into multiple .cc
files to allow parallelization of the compilation.

While this technically doesn't really solve the problem of reducing
compile load (like some pimpl techniques would) and instead rather
increases the total compilation load it works well in practice because
dev and CI machines have high core counts and as described above the
final phase of compilation is effectively single threaded today.

Comparing something like

```
bazel build --config=release //src/v/redpanda --copt -mllvm --copt -inline-threshold=230
```

on m8gd.12xl we see:

Previous:

```
INFO: Analyzed target //src/v/redpanda:redpanda (1 packages loaded, 68528 targets configured).
INFO: Found 1 target...
Target //src/v/redpanda:redpanda up-to-date:
  bazel-bin/src/v/redpanda/redpanda
INFO: Elapsed time: 672.215s, Critical Path: 397.27s
INFO: 1804 processes: 6485 action cache hit, 15 internal, 1789 processwrapper-sandbox.
INFO: Build completed successfully, 1804 total actions

real    11m12.329s
```

With these changes:

```
INFO: Analyzed target //src/v/redpanda:redpanda (1 packages loaded, 68535 targets configured).
...
Target //src/v/redpanda:redpanda up-to-date:
  bazel-bin/src/v/redpanda/redpanda
INFO: Elapsed time: 569.678s, Critical Path: 251.68s
INFO: 1810 processes: 6486 action cache hit, 14 internal, 1796 processwrapper-sandbox.
INFO: Build completed successfully, 1810 total actions

real    9m29.778s
```

So it's quite a measurable improvement.

Similarly locally when "changing" application.h to trigger a application
tree recompile:

```
stephan@rp:/build/redpanda$ time bazel build --config=sanitizer //src/v/redpanda:application
INFO: Invocation ID: b8bf72b8-0d5c-4c6b-b571-a71270475a7c
INFO: Analyzed target //src/v/redpanda:application (0 packages loaded, 0 targets configured).
INFO: Found 1 target...
Target //src/v/redpanda:application up-to-date:
  bazel-bin/src/v/redpanda/libapplication.a
  bazel-bin/src/v/redpanda/libapplication.so
INFO: Elapsed time: 225.546s, Critical Path: 225.18s
INFO: 3 processes: 18 action cache hit, 2 internal, 1 linux-sandbox.
INFO: Build completed successfully, 3 total actions
```

vs.

```
stephan@rp:/build/redpanda$ time bazel build --config=sanitizer //src/v/redpanda:application
INFO: Invocation ID: cacb6dc5-a356-405a-914c-95b3d65e45ac
INFO: Analyzed target //src/v/redpanda:application (1 packages loaded, 15 targets configured).
INFO: Found 1 target...
Target //src/v/redpanda:application up-to-date:
  bazel-bin/src/v/redpanda/libapplication.a
  bazel-bin/src/v/redpanda/libapplication.so
INFO: Elapsed time: 116.513s, Critical Path: 116.09s
INFO: 14 processes: 19 action cache hit, 3 internal, 11 linux-sandbox.
INFO: Build completed successfully, 14 total actions
```
